feat(paths,markers): Add more watch options
This commit adds the ability to a) modify the watch options of paths
in a similar manner to geojson and markers and b) to use a 'watchCollection'
strategy instead of 'watch' for these collections. Specifically, if the 'doWatch'
property of xWatchOptions is set to 'collection', then 'watchCollection' is
used instead of 'watch'. For long paths that need to be updated often, the
'collection' strategy, combined with watching individual path changes, works is
much more effective than deep watching the entire paths object or even
individual paths. An example of this setup would be:

        {
            doWatch: 'collection',
            isDeep: false,
            individual: {
                doWatch: true,
                isDeep: false
            }
        }

It's worth noting that before this commit, watching paths was analogous to
having the paths watch options set to:

        {
            doWatch: 'collection',
            isDeep: true,
            individual: {
                doWatch: shouldWatch,
                isDeep: true
            }
        }

where shouldWatch is whatever watch-paths was set to (by default true).

I also fixed an error in markers.js where oldModel was always set to
undefined, even if the oldModel existed.