docs(markers-attribute): Update links to examples
For example, the markers-angular-template-example.html link was stale
because:

* eaa051c3 (Removed old examples, 2015-04-17, #697) removed the
  unprefixed version.
* 2c40797e (new versions of examples, 2015-04-17, #697) added an 0514
  prefix.
* 726d37ce (fix(examples) Fix ESRI legend examples, 2015-09-18, #949)
  changed the prefix to 0515.

The only example that did not retain its name across the eaa051c3 /
2c40797e transition was markers-update-example.html.  I'm assuming
that the new version of that example is
0504-markers-popup-example.html, because it is the only new example
HTML which contains leafletDirectiveMarker.popupopen.

This commit just updates the links.  I have not checked the
surrounding text to see if it is also stale.